+++ src/Fragment.spec.ts
@@ -44,13 +44,17 @@ describe( module( "Carbon/Fragment" ), ():void => {
 			expect( Fragment.Factory.hasClassProperties ).toBeDefined();
 			expect( Utils.isFunction( Fragment.Factory.hasClassProperties ) ).toBe( true );
 
-			let resource:Object = undefined;
-			expect( Fragment.Factory.hasClassProperties( resource ) ).toBe( false );
-			resource = {};
+			let resource:any = undefined;
 			expect( Fragment.Factory.hasClassProperties( resource ) ).toBe( false );
 
-			resource["document"] = null;
+			resource = {
+				document: null
+			};
 			expect( Fragment.Factory.hasClassProperties( resource ) ).toBe( true );
+
+			delete resource.document;
+			expect( Fragment.Factory.hasClassProperties( resource ) ).toBe( false );
+			resource.document = null;
 		});
 
 		let document:Document.Class;
@@ -216,35 +220,6 @@ describe( module( "Carbon/Fragment" ), ():void => {
 				expect( anotherFragment[ "property" ] ).toBeUndefined();
 			});
 
-		});
-
-	});
-
-	describe( clazz(
-		"Carbon.Fragment.Util",
-		"Class with useful options for Fragment objects"
-	), ():void => {
-
-		it( isDefined(), ():void => {
-			expect( Fragment.Util ).toBeDefined();
-			expect( Utils.isFunction( Fragment.Util ) ).toBe( true );
-		});
-
-		it( hasMethod(
-			STATIC,
-			"generateID",
-			"Returns an ID for a BlankNode using an universally unique identifier (UUID)."
-		), ():void => {
-			let id1:string;
-			let id2:string;
-
-			id1 = Fragment.Util.generateID();
-			expect( URI.Util.isBNodeID( id1 ) ).toBe( true );
-
-			id2 = Fragment.Util.generateID();
-			expect( URI.Util.isBNodeID( id2 ) ).toBe( true );
-
-			expect( id1 ).not.toEqual( id2 );
 		});
 
 	});