+++ src/LDP.ts
@@ -6,6 +6,8 @@ import * as IndirectContainer from "./LDP/IndirectContainer";
 import * as PersistedContainer from "./LDP/PersistedContainer";
 import * as RDFSource from "./LDP/RDFSource";
 import * as RemoveMemberAction from "./LDP/RemoveMemberAction";
+import * as ResponseMetadata from "./LDP/ResponseMetadata";
+import * as ResourceMetadata from "./LDP/ResourceMetadata";
 
 export {
 	AddMemberAction,
@@ -16,4 +18,6 @@ export {
 	PersistedContainer,
 	RDFSource,
 	RemoveMemberAction,
+	ResponseMetadata,
+	ResourceMetadata,
 };