+++ src/LDP/PersistedContainer.ts
@@ -3,6 +3,7 @@ import * as HTTP from "./../HTTP";
 import * as PersistedDocument from "./../PersistedDocument";
 import * as Pointer from "./../Pointer";
 import * as Utils from "./../Utils";
+import * as RetrievalPreferences from "./../RetrievalPreferences";
 
 export interface Class extends PersistedDocument.Class {
 	addMember( member:Pointer.Class ): Promise<HTTP.Response.Class>;
@@ -15,9 +16,14 @@ export interface Class extends PersistedDocument.Class {
 	createChild( object:Object ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 	createChild():Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 
-	getChildren():Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+	listChildren():Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
 
-	getMembers( includeNonReadable?:boolean ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+	getChildren( retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+
+	listMembers( includeNonReadable?:boolean ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+
+	getMembers( includeNonReadable?:boolean, retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+	getMembers( retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
 
 	removeMember( member:Pointer.Class ): Promise<HTTP.Response.Class>;
 	removeMember( memberURI:string ): Promise<HTTP.Response.Class>;
@@ -59,13 +65,24 @@ function createChild( slugOrObject?:any, object?:Object ):Promise<[ Pointer.Clas
 	}
 }
 
-function getChildren():Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
+function listChildren():Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
 	let that:PersistedDocument.Class = <PersistedDocument.Class> this;
-	return this._documents.getChildren( that.id );
+	return this._documents.listChildren( that.id );
 }
 
-function getMembers( includeNonReadable:boolean = true ):Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
-	return this._documents.getMembers( this.id, includeNonReadable );
+function getChildren( retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
+	let that:PersistedDocument.Class = <PersistedDocument.Class> this;
+	return this._documents.getChildren( that.id, retrievalPreferences );
+}
+
+function listMembers( includeNonReadable:boolean = true ):Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
+	return this._documents.listMembers( this.id, includeNonReadable );
+}
+
+function getMembers( includeNonReadable:boolean, retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+function getMembers( retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+function getMembers( nonReadRetPref:boolean = true, retrievalPreferences?:RetrievalPreferences.Class ):Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
+	return this._documents.getMembers( this.id, nonReadRetPref, retrievalPreferences );
 }
 
 function removeMember( member:Pointer.Class ): Promise<HTTP.Response.Class>;
@@ -102,10 +119,13 @@ function upload( slugOrData:any, data:any = null ):Promise<[ Pointer.Class, HTTP
 
 	export class Factory {
 	static hasClassProperties( document:Document.Class ):boolean {
-		return Utils.hasFunction( document, "addMember" )
+		return Utils.isObject( document )
+			&& Utils.hasFunction( document, "addMember" )
 			&& Utils.hasFunction( document, "addMembers" )
 			&& Utils.hasFunction( document, "createChild" )
+			&& Utils.hasFunction( document, "listChildren" )
 			&& Utils.hasFunction( document, "getChildren" )
+			&& Utils.hasFunction( document, "listMembers" )
 			&& Utils.hasFunction( document, "getMembers" )
 			&& Utils.hasFunction( document, "removeMember" )
 			&& Utils.hasFunction( document, "removeMembers" )
@@ -135,11 +155,23 @@ function upload( slugOrData:any, data:any = null ):Promise<[ Pointer.Class, HTTP
 				configurable: true,
 				value: createChild,
 			},
+			"listChildren": {
+				writable: false,
+				enumerable: false,
+				configurable: true,
+				value: listChildren,
+			},
 			"getChildren": {
 				writable: false,
 				enumerable: false,
 				configurable: true,
 				value: getChildren,
+			},
+			"listMembers": {
+				writable: false,
+				enumerable: false,
+				configurable: true,
+				value: listMembers,
 			},
 			"getMembers": {
 				writable: false,